+++ CHANGELOG.md
@@ -11,8 +11,10 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
-### 2025-12-XX
+### 2025-12-08
   * Improved get drawing_def edit bound computation for container without faces
+  * Added the centred option to Stretch Tool
+  * Updated PackingSolver dependency
 ### 2025-12-06
   * Various improvements
 ### 2025-11-30